requirements.txt
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
astroid==3.3.11
asttokens==3.0.0
async-timeout==5.0.1
audeer==2.3.1
audformat==1.3.2
audinterface==1.3.1
audiofile==1.5.1
audmath==1.4.4
audobject==0.7.12
audresample==1.3.5
autopep8==2.0.4
black==25.1.0
blessed==1.25.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
contourpy==1.2.1
cycler==0.12.1
datasets==4.0.0
deepmultilingualpunctuation==1.0.1
dill==0.3.8
docstring-to-markdown==0.17
einops==0.8.0
face-alignment==1.4.1
filelock==3.20.0
flake8==7.1.2
fonttools==4.60.1
frozenlist==1.8.0
gdown==5.2.0
gpustat==1.1.1
hf-xet==1.1.10
huggingface-hub==0.35.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipdb==0.13.13
iso3166==2.1.1
iso639-lang==2.6.3
isort==6.0.1
jedi==0.17.2
Jinja2==3.1.6
joblib==1.5.2
jupyterlab-lsp==5.2.0
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.7.5
mccabe==0.7.0
mdurl==0.1.2
mmcv==2.2.0
mmengine==0.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
munkres==1.1.4
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
opencv-python==4.11.0.86
opensmile==2.6.0
oyaml==1.0
pandas==2.3.0
parso==0.7.1
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.4.0
propcache==0.4.1
pyarrow==22.0.0
pycodestyle==2.12.1
pycparser==2.23
pydocstyle==6.3.0
pyflakes==3.2.0
pylint==3.3.8
pyparsing==3.2.5
python-jsonrpc-server==0.4.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.13.1
pytoolconfig==1.3.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.22
requests==2.32.5
resampy==0.4.3
rich==14.2.0
rope==1.14.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
six==1.16.0
snowballstemmer==3.0.1
soundfile==0.13.1
srt==3.5.3
sympy==1.13.1
termcolor==3.2.0
theme-darcula==4.0.0
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
transformers==4.57.1
triton==3.1.0
tzdata==2025.2
ujson==5.11.0
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
whatthepatch==1.0.7
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.23.0
